requirements.txt
flask==2.2.3
tensorflow>=2.19.0
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.3
scikit-image==0.20.0
gdown==5.1.0
imageio==2.31.1
tensorflow-addons==0.21.0